Add Remove Commands

+++ bot.js
@@ -36,10 +36,14 @@ function handleCourseCommands(msg) {
 }
 
 function handleRoleCommands(msg) {
-  const roleCommand = roleCommands.find(roleCommand => {
-    return msg.content === `${BOT_PREFIX}${roleCommand.command}`
-  })
+  handleAddRoleCommands(msg)
+  handleRemoveRoleCommands(msg)
+}
 
+function handleAddRoleCommands(msg) {
+  const roleCommand = roleCommands.find(roleCommand => {
+    return msg.content === `${BOT_PREFIX}add-${roleCommand.command}-role`
+  })
   if (roleCommand == null) return
 
   const role = msg.guild.roles.cache.get(roleCommand.roleId)
@@ -48,7 +52,27 @@ function handleRoleCommands(msg) {
   Promise.all([
     msg.member.roles.add(role),
     msg.member.send(
-      `You will now be notified when new ${roleCommand.type} are posted.`
+      `You will now be notified when new ${roleCommand?.type} are posted.`
+    ),
+    msg.delete(),
+  ]).catch(e => {
+    console.error(e)
+  })
+}
+
+function handleRemoveRoleCommands(msg) {
+  const roleCommand = roleCommands.find(roleCommand => {
+    return msg.content === `${BOT_PREFIX}remove-${roleCommand.command}-role`
+  })
+  if (roleCommand == null) return
+
+  const role = msg.guild.roles.cache.get(roleCommand.roleId)
+  if (role == null) return
+
+  Promise.all([
+    msg.member.roles.remove(role),
+    msg.member.send(
+      `You will no longer be notified when new ${roleCommand?.type} are posted.`
     ),
     msg.delete(),
   ]).catch(e => {
